CORE-4720 Fixes for duplicate App listings and counts.
Removed the Group ID and name from the get-app-listing-base-query and added a "DISTINCT" modifier to prevent the query from returning more than 1 result per App.
Refactored the template_group_template join out of the get-app-listing-base-query and get-app-count-base-query functions, and into the 'where' clauses used with these base queries.
Refactored the 'count' in get-app-count-base-query to include a "DISTINCT" modifier on app IDs to prevent the query from counting an App more than once.